# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: UseCase Manager
  description: AI-powered use case extraction from workshop transcripts
  version: 0.1.0
paths:
...
  /companies:
    get:
      tags:
      - companies
      summary: List Companies
      description: List all companies with optional filters
      operationId: list_companies_companies_get
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Page Size
      - name: industry_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: "null"
          title: Industry Id
      - name: q
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: "null"
          title: Q
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response List Companies Companies Get
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                properties:
                  detail:
                    items:
                      properties:
                        loc:
                          items:
                            anyOf:
                            - type: string
                            - type: integer
                          type: array
                          title: Location
                        msg:
                          type: string
                          title: Message
                        type:
                          type: string
                          title: Error Type
                      type: object
                      required:
                      - loc
                      - msg
                      - type
                      title: ValidationError
                    type: array
                    title: Detail
                type: object
                title: HTTPValidationError
    post:
      tags:
      - companies
      summary: Create Company
      description: Create a new company
      operationId: create_company_companies_post
      parameters:
      - name: email
        in: query
        required: true
        schema:
          type: string
          title: Email
      - name: full_name
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: "null"
          title: Full Name
      - name: id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: Id
      - name: role
        in: query
        required: true
        schema:
          type: string
          enum:
          - reader
          - maintainer
          - admin
          title: UserRole
      - name: is_active
        in: query
        required: true
        schema:
          type: boolean
          title: Is Active
      - name: is_verified
        in: query
        required: true
        schema:
          type: boolean
          title: Is Verified
      - name: created_at
        in: query
        required: true
        schema:
          type: string
          format: date-time
          title: Created At
      - name: updated_at
        in: query
        required: true
        schema:
          type: string
          format: date-time
          title: Updated At
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                name:
                  type: string
                  title: Name
                industry_id:
                  type: string
                  format: uuid
                  title: Industry Id
                description:
                  anyOf:
                  - type: string
                  - type: "null"
                  title: Description
                website:
                  anyOf:
                  - type: string
                  - type: "null"
                  title: Website
              type: object
              required:
              - name
              - industry_id
              title: CompanyCreate
      responses:
        "201":
          description: Successful Response
          content:
            application/json:
              schema:
                properties:
                  name:
                    type: string
                    title: Name
                  industry_id:
                    type: string
                    format: uuid
                    title: Industry Id
                  description:
                    anyOf:
                    - type: string
                    - type: "null"
                    title: Description
                  website:
                    anyOf:
                    - type: string
                    - type: "null"
                    title: Website
                  id:
                    type: string
                    format: uuid
                    title: Id
                  created_at:
                    type: string
                    format: date-time
                    title: Created At
                  updated_at:
                    type: string
                    format: date-time
                    title: Updated At
                type: object
                required:
                - name
                - industry_id
                - id
                - created_at
                - updated_at
                title: CompanyResponse
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                properties:
                  detail:
                    items:
                      properties:
                        loc:
                          items:
                            anyOf:
                            - type: string
                            - type: integer
                          type: array
                          title: Location
                        msg:
                          type: string
                          title: Message
                        type:
                          type: string
                          title: Error Type
                      type: object
                      required:
                      - loc
                      - msg
                      - type
                      title: ValidationError
                    type: array
                    title: Detail
                type: object
                title: HTTPValidationError
  /companies/{company_id}:
    get:
      tags:
      - companies
      summary: Get Company
      description: Get a specific company
      operationId: get_company_companies__company_id__get
      parameters:
      - name: company_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Company Id
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                properties:
                  name:
                    type: string
                    title: Name
                  industry_id:
                    type: string
                    format: uuid
                    title: Industry Id
                  description:
                    anyOf:
                    - type: string
                    - type: "null"
                    title: Description
                  website:
                    anyOf:
                    - type: string
                    - type: "null"
                    title: Website
                  id:
                    type: string
                    format: uuid
                    title: Id
                  created_at:
                    type: string
                    format: date-time
                    title: Created At
                  updated_at:
                    type: string
                    format: date-time
                    title: Updated At
                type: object
                required:
                - name
                - industry_id
                - id
                - created_at
                - updated_at
                title: CompanyResponse
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                properties:
                  detail:
                    items:
                      properties:
                        loc:
                          items:
                            anyOf:
                            - type: string
                            - type: integer
                          type: array
                          title: Location
                        msg:
                          type: string
                          title: Message
                        type:
                          type: string
                          title: Error Type
                      type: object
                      required:
                      - loc
                      - msg
                      - type
                      title: ValidationError
                    type: array
                    title: Detail
                type: object
                title: HTTPValidationError
    patch:
      tags:
      - companies
      summary: Update Company
      description: Update a company
      operationId: update_company_companies__company_id__patch
      parameters:
      - name: company_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Company Id
      - name: email
        in: query
        required: true
        schema:
          type: string
          title: Email
      - name: full_name
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: "null"
          title: Full Name
      - name: id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: Id
      - name: role
        in: query
        required: true
        schema:
          type: string
          enum:
          - reader
          - maintainer
          - admin
          title: UserRole
      - name: is_active
        in: query
        required: true
        schema:
          type: boolean
          title: Is Active
      - name: is_verified
        in: query
        required: true
        schema:
          type: boolean
          title: Is Verified
      - name: created_at
        in: query
        required: true
        schema:
          type: string
          format: date-time
          title: Created At
      - name: updated_at
        in: query
        required: true
        schema:
          type: string
          format: date-time
          title: Updated At
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                name:
                  anyOf:
                  - type: string
                  - type: "null"
                  title: Name
                industry_id:
                  anyOf:
                  - type: string
                    format: uuid
                  - type: "null"
                  title: Industry Id
                description:
                  anyOf:
                  - type: string
                  - type: "null"
                  title: Description
                website:
                  anyOf:
                  - type: string
                  - type: "null"
                  title: Website
              type: object
              title: CompanyUpdate
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                properties:
                  name:
                    type: string
                    title: Name
                  industry_id:
                    type: string
                    format: uuid
                    title: Industry Id
                  description:
                    anyOf:
                    - type: string
                    - type: "null"
                    title: Description
                  website:
                    anyOf:
                    - type: string
                    - type: "null"
                    title: Website
                  id:
                    type: string
                    format: uuid
                    title: Id
                  created_at:
                    type: string
                    format: date-time
                    title: Created At
                  updated_at:
                    type: string
                    format: date-time
                    title: Updated At
                type: object
                required:
                - name
                - industry_id
                - id
                - created_at
                - updated_at
                title: CompanyResponse
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                properties:
                  detail:
                    items:
                      properties:
                        loc:
                          items:
                            anyOf:
                            - type: string
                            - type: integer
                          type: array
                          title: Location
                        msg:
                          type: string
                          title: Message
                        type:
                          type: string
                          title: Error Type
                      type: object
                      required:
                      - loc
                      - msg
                      - type
                      title: ValidationError
                    type: array
                    title: Detail
                type: object
                title: HTTPValidationError
    delete:
      tags:
      - companies
      summary: Delete Company
      description: Delete a company
      operationId: delete_company_companies__company_id__delete
      parameters:
      - name: company_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Company Id
      - name: email
        in: query
        required: true
        schema:
          type: string
          title: Email
      - name: full_name
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: "null"
          title: Full Name
      - name: id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: Id
      - name: role
        in: query
        required: true
        schema:
          type: string
          enum:
          - reader
          - maintainer
          - admin
          title: UserRole
      - name: is_active
        in: query
        required: true
        schema:
          type: boolean
          title: Is Active
      - name: is_verified
        in: query
        required: true
        schema:
          type: boolean
          title: Is Verified
      - name: created_at
        in: query
        required: true
        schema:
          type: string
          format: date-time
          title: Created At
      - name: updated_at
        in: query
        required: true
        schema:
          type: string
          format: date-time
          title: Updated At
      responses:
        "204":
          description: Successful Response
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                properties:
                  detail:
                    items:
                      properties:
                        loc:
                          items:
                            anyOf:
                            - type: string
                            - type: integer
                          type: array
                          title: Location
                        msg:
                          type: string
                          title: Message
                        type:
                          type: string
                          title: Error Type
                      type: object
                      required:
                      - loc
                      - msg
                      - type
                      title: ValidationError
                    type: array
                    title: Detail
                type: object
                title: HTTPValidationError
...
components:
...
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: auth/jwt/login